removed voting system to change later
wasted hours trying to replicate ui across players similar to a lobby. So I gave up and are going to create one which the players will have to run to torches in the corner of the rooms. Each torch represents a players position (1-4), which ever position they are in will determine a buff. Only 1 player can interact with a torch to ensure that multiple of the same position doesn't occur.